tests: Bluetooth: Audio: Add common stream_started_cb
Many profiles implemented their own nearly identical
stream_started_cb. This meant not only duplicated code, but
also callbacks that not always did the same expected behavior.

Added a new bap_common_stream_started_cb that can be used directly
as a callback, or be called by the profiles own callback if the
profile test needs any additional checks.

This also removed unused audio_test_stream.tx_sdu_size as well
as a `static bool last_failed;`
