+++ tests/bsim/bluetooth/audio/src/bap_common.h
@@ -121,4 +121,5 @@ bool bap_stream_is_streaming(const struct bt_bap_stream *bap_stream);
 bool cap_stream_is_streaming(const struct bt_cap_stream *cap_stream);
 bool audio_test_stream_is_streaming(const struct audio_test_stream *test_stream);
 void bap_unicast_stream_disconnected_cb(struct bt_bap_stream *stream, uint8_t reason);
+void bap_common_stream_started_cb(struct bt_bap_stream *stream);
 #endif /* ZEPHYR_TEST_BSIM_BT_AUDIO_TEST_COMMON_ */
