+++ tests/bsim/bluetooth/audio/src/bap_unicast_server_test.c
@@ -156,15 +156,11 @@ static int lc3_reconfig(struct bt_bap_stream *stream, enum bt_audio_dir dir,
 static int lc3_qos(struct bt_bap_stream *stream, const struct bt_bap_qos_cfg *qos,
 		   struct bt_bap_ascs_rsp *rsp)
 {
-	struct audio_test_stream *test_stream = audio_test_stream_from_bap_stream(stream);
-
 	ARG_UNUSED(rsp);
 
 	LOG_INF("QoS: stream %p qos %p", stream, qos);
 
 	print_qos(qos);
-
-	test_stream->tx_sdu_size = qos->sdu;
 
 	return 0;
 }
@@ -309,17 +305,7 @@ static void stream_enabled_cb(struct bt_bap_stream *stream)
 
 static void stream_started_cb(struct bt_bap_stream *stream)
 {
-	LOG_INF("Started: stream %p", stream);
-
-	if (bap_stream_tx_can_send(stream)) {
-		int err;
-
-		err = bap_stream_tx_register(stream);
-		if (err != 0) {
-			FAIL("Failed to register stream %p for TX: %d\n", stream, err);
-			return;
-		}
-	}
+	bap_common_stream_started_cb(stream);
 
 	SET_FLAG(flag_stream_started);
 }
