+++ tests/bsim/bluetooth/audio/src/cap_handover_peripheral_test.c
@@ -243,15 +243,7 @@ static void stream_enabled_cb(struct bt_bap_stream *stream)
 
 static void stream_started_cb(struct bt_bap_stream *stream)
 {
-	struct audio_test_stream *test_stream = audio_test_stream_from_bap_stream(stream);
-
-	memset(&test_stream->last_info, 0, sizeof(test_stream->last_info));
-	test_stream->rx_cnt = 0U;
-	test_stream->valid_rx_cnt = 0U;
-	test_stream->seq_num = 0U;
-	test_stream->tx_cnt = 0U;
-
-	LOG_DBG("Started stream %p", stream);
+	bap_common_stream_started_cb(stream);
 
 	SET_FLAG(flag_stream_started);
 }
